Initial commit
This package allows the user to Download data directly from Yahoo finance. It offers tools to analyze and visualize the stocks prices and returns in order to gain a deeper statistical understanding. The user has also the possibility to form a portfolio of stocks in which to run the same analysis and test different back-testing strategies.